refactor: Remove types/internal.py and consolidate into solver_types.py
**Problem**: The types/internal.py module contained types that had already
been moved to other locations, leading to:
- Broken imports (5 files importing non-existent types)
- Duplicated type definitions
- Confusing module organization
- Reference to non-existent `_internal/` directory in CLAUDE.md

**Solution**: Complete consolidation of type definitions:

1. **Fixed broken imports** (5 files):
   - JAXArray: amr_1d.py, amr_quadtree_2d.py, jax_utils.py
   - MultiIndexTuple: hjb_gfdm.py
   - SolverReturnTuple: base_mfg.py
   - All now import from `types.solver_types` instead of `types.internal`

2. **Migrated unique content** to solver_types.py:
   - HamiltonianFunction, LagrangianFunction, DensityFunction, ValueFunction
   - NewtonSolver, LinearSolver protocols
   - SolverError, ConvergenceError, ConfigurationError exceptions
   - FlexibleInput type

3. **Deleted obsolete module**:
   - Removed mfg_pde/types/internal.py entirely
   - Updated types/__init__.py comments to reflect new structure

4. **Updated documentation**:
   - Removed `_internal/` reference from CLAUDE.md
   - Added missing directories (hooks/, types/) to package structure
   - Accurate reflection of actual repository organization

**Impact**:
- All type imports now work correctly
- No duplication of type definitions
- Clearer organization (solver_types.py is comprehensive)
- Documentation matches reality

**Files Changed**:
- Deleted: mfg_pde/types/internal.py
- Updated: 5 import statements across codebase
- Enhanced: mfg_pde/types/solver_types.py (added Protocol import, moved types)
- Updated: mfg_pde/types/__init__.py, CLAUDE.md

**Testing**: All imports verified, pre-commit hooks passed
