+++ mfg_pde/types/__init__.py
@@ -39,8 +39,6 @@ __all__ = [
 # Array types available via explicit import:
 # from mfg_pde.types.arrays import SpatialArray, TemporalArray, ParticleArray, ...
 
-# Solver types available via explicit import:
+# Solver types and advanced types available via explicit import:
 # from mfg_pde.types.solver_types import JAXSolverReturn, ComplexSolverState, MetadataDict, ...
-
-# Internal types available via explicit import:
-# from mfg_pde.types.internal import HamiltonianFunction, ...
+# from mfg_pde.types.solver_types import HamiltonianFunction, NewtonSolver, SolverError, ...
